+++ src/kmacho/load_commands.py
@@ -86,7 +86,7 @@ class SegmentLoadCommand(LoadCommand):
         command_type = segment_command_64 if is_64 else segment_command
         section_type = section_64 if is_64 else section
         cmd = 0x19 if is_64 else 0x1
-        # 'cmd', 'cmdsize', 'segname', 'vmaddr', 'vmsize', 'fileoff', 'filesize', 'maxprot', 'initprot', 'nsects', 'flags'
+
         cmdsize = command_type.SIZE
         cmdsize += (len(sections) * section_type.SIZE)
 
@@ -128,4 +128,4 @@ class SegmentLoadCommand(LoadCommand):
 
         self.sections = {}
 
-
+# TODO: Constructable wrapper for dylinker_command, build_version_command
