+++ src/ktool/__init__.py
@@ -1,7 +1,7 @@
 from ktool.ktool import load_image, load_objc_metadata, generate_headers, generate_text_based_stub, load_macho_file, macho_verify, reload_image, macho_combine
 
 from ktool.objc import ObjCImage
-from ktool.dyld import Dyld, LD64, Image
+from ktool.loader import MachOImageLoader, Image
 from ktool.macho import Slice, MachOFile, MachOFileType, Segment, Section
 try:
     from ktool.headers import HeaderGenerator, Header
